fix CI

+++ tools/leech/main.cc
@@ -39,7 +39,7 @@ int main(int argc, char **argv) try {
     timer::Timer timer;
     vm.run();
     auto time = timer.elapsed_mcs();
-    std::cout << "Time: " << time * 1e-3 << " ms" << std::endl;
+    std::cout << "Time: " << static_cast<double>(time) * 1e-3 << " ms" << std::endl;
   }
 } catch (const std::exception &e) {
   spdlog::error(e.what());
